+++ src/stdlib/map/map.bpf.c
@@ -7,11 +7,8 @@
 struct bpf_map;
 extern __s64 bpf_map_sum_elem_count(const struct bpf_map *map) __ksym __weak;
 
-_Bool __has_key(void *map, void *key) {
-    if (bpf_map_lookup_elem(map, key) == NULL) {
-        return 0;
-    }
-    return 1;
+void * __lookup_elem(void *map, void *key) {
+    return bpf_map_lookup_elem(map, key);
 }
 
 long __delete(void *map, void *key) {